Allow `persistgraphql` to exclude certain paths
Summary:
In some cases, we may want `persistgraphql` to skip over certain files
or directories that we don't want to look for GraphQL queries in. For
example, in the `mobile` repo, we want to look for queries in
`react-native` but not search `react-native/node_modules`, to avoid picking up
queries not in our code, speed up the command, and prevent errors from
queries we don't care about. This change adds the ability to specify
paths to exclude via adding an argument to the command: `--exclude=path1,path2,path3`

I also normalize the paths so that there aren't unexpected bugs from
inconsistent path inputs. For example, using `directory/` as the input
path and `directory/path/` as the excluded path would break without normalization,
because as we traverse `directory`, the input path becomes
`directory//path` which wouldn't match the excluded path.

Note: the only real changes are in `src/ExtractGQL.ts`. The other file changes are just changes in the resulting compiled files.

Test Plan:
`npm run compile`
Try excluding some file or directory:
`bin/persistgraphql <path to webapp root>/javascript --add_typename --js
--extension=jsx,js --exclude=<path to webapp root>/javascript/assignments-package`
- You should see `Excluding path __` in the output
- The resulting `extracted_queries.json` should not contain queries such
  as `archiveAssignment`, `CoachAssignments`, `createAssignment`,
  `UserAssignments`.

This should also work on a specific path, or multiple comma-separated
paths.

Reviewers: dhruv, john, alex

Reviewed By: dhruv

Differential Revision: https://phabricator.khanacademy.org/D38878