Export more of the API for use by upload_graphql_safelist.js. (#2)
## Summary:
These are probably API components that the author considered internal,
but given that this code hasn't changed in over 3 years I feel the API
is probably pretty set at this point.

The motivation for this is to allow us to run persistgraphql, as part
of our safelist-upload process, without having to shell out to run it
as a binary.  Secondary benefits mean not having to involve temporary
files, which the binary API requires.

Issue: none

## Test plan:
Will try this in webapp!

Author: csilvers

Reviewers: jeresig

Required Reviewers:

Approved By: jeresig

Checks:

Pull Request URL: https://github.com/Khan/persistgraphql/pull/2